fix conversion warning

+++ grunk/core/Vec.hpp
@@ -29,7 +29,7 @@ public:
     {
         std::vector<T> v;
         v.reserve(this->parents.size());
-        for (size_t i=0; i<this->parents.size(); ++i) {
+        for (int i=0; i<this->parents.size(); ++i) {
             v.push_back(this->template arg<T>(i).value());
         }
         if (auto r =  this->template res<0>(); r) {
